test: Update expected e2e test outputs
